andrews cosmetic fixes

+++ lin_guider/src/settings.cpp
@@ -75,8 +75,8 @@ settings::settings( lin_guider *parent,
 	for( int i = 0;i < cnt;i++ )
 		ui.comboBox_GuiderAlgorithm->addItem( QString(lg_math::alg_desc_list[i].desc), lg_math::alg_desc_list[i].type );
 
-	double zoom_max = ZOOM_MAX * 100;
-	double zoom_step = ZOOM_STEP * 100;
+	float zoom_max = ZOOM_MAX * 100;
+	float zoom_step = ZOOM_STEP * 100;
 	for( float i = 0; i < zoom_max; i += zoom_step )
 		ui.comboBox_VP_scale->addItem( QString().setNum((i + zoom_step)), (i + zoom_step) / 100 );
